Move cache size & clear logic to ChannelList component
